feat(frontend): analytics events, SEO assets/structured-data/FAQ, retire /chat UI
Analytics (Umami):
- Add typed trackEvent() wrapper (src/lib/analytics.ts); cookieless, non-PII payloads
- Instrument funnel/adoption/outbound/friction events: sign-in, overlay create,
  source add, OBS-URL copy, theme apply, TTS/sound enable, share accept,
  YT stream-strategy, outbound + CTA clicks

SEO:
- Add favicon (icon.svg), apple-icon, and web manifest (theme-color #0f0f13)
- Add Organization + WebSite JSON-LD (layout) via a JsonLd helper
- Split homepage into a server wrapper (canonical + SoftwareApplication + FAQPage
  JSON-LD) over HomeClient; add per-page canonicals to the legal pages
- Add a visible landing FAQ (FaqSection + shared faq.ts) with matching FAQPage schema
  (copy pending owner review)

Retire deprecated /chat viewer-chat UI:
- Remove /chat/[streamer] and VIEWER_CHAT_SETUP.md; prune now-unused viewer plumbing
- Keep /chat/auth-success & /chat/auth-error (load-bearing OAuth catchers used by the
  browser extension + auth-service redirects); slim auth-success's dead fallback
- robots.ts: move /chat/ from allow to disallow
