--- VIEWER_CHAT_SETUP.md
@@ -1,173 +1,0 @@
-# Viewer Chat Feature - Configuration Setup
-
-**Status**: ✅ Code Complete | ⚠️ Configuration Required
-**Date**: 2025-12-19
-
-## 🎉 Implementation Complete
-
-All code has been implemented, tested, and deployed:
-- ✅ Backend API endpoints (auth-service)
-- ✅ Frontend UI (chat pages)
-- ✅ Database migration applied
-- ✅ CI/CD deployment successful
-
-## ⚠️ Required Configuration
-
-### Twitch Developer Console Setup
-
-The viewer OAuth feature requires adding a new redirect URI to the Twitch application.
-
-**Steps:**
-
-1. **Go to Twitch Developer Console**
-   - Visit: https://dev.twitch.tv/console/apps
-   - Log in with your Twitch account
-   - Find your All-Chat application
-
-2. **Add Viewer Callback URI**
-   - Click "Manage" on your application
-   - Scroll to "OAuth Redirect URLs"
-   - Add new URL: `https://allch.at/api/v1/auth/viewer/twitch/callback`
-   - Click "Add" and then "Save"
-
-**Current Redirect URIs** (should include):
-- `https://allch.at/api/v1/auth/twitch/callback` (existing - for streamer auth)
-- `https://allch.at/api/v1/auth/viewer/twitch/callback` (NEW - for viewer auth)
-
-### Why This Is Needed
-
-The viewer OAuth flow uses a different callback URL than the streamer OAuth flow to:
-- Keep viewer and streamer authentication separate
-- Allow different scopes (viewer needs `user:write:chat`)
-- Prevent token confusion between the two auth contexts
-
-## 📊 Test Results
-
-### ✅ Successful Tests
-
-1. **UI Rendering** ✅
-   - Chat page loads at `/chat/caesarlp`
-   - Displays streamer name and active platforms
-   - Shows login button
-   - Proper styling with Tailwind CSS
-
-2. **OAuth Initiation** ✅
-   - Login button redirects to Twitch OAuth
-   - Correct parameters in OAuth URL:
-     - client_id: ✓
-     - redirect_uri: ✓
-     - scope: `user:write:chat` ✓
-     - state: ✓ (CSRF protection)
-
-3. **API Endpoints** ✅
-   - Streamer info: `GET /api/v1/auth/streamers/caesarlp` - Working
-   - OAuth login: `GET /api/v1/auth/viewer/twitch/login` - Working
-
-### ⚠️ Configuration Blocker
-
-**Issue**: Twitch OAuth returns `redirect_mismatch` error
-**Reason**: Redirect URI not registered in Twitch app
-**Fix**: Add `https://allch.at/api/v1/auth/viewer/twitch/callback` to Twitch app settings
-**Impact**: Blocks OAuth completion until configured
-
-## 🚀 Feature Capabilities (Once Configured)
-
-1. **Viewer Login**
-   - Visit `/chat/{streamer_username}`
-   - Click "Login with Twitch"
-   - Authorize with Twitch (grants `user:write:chat` scope)
-   - Redirected back to chat page (authenticated)
-
-2. **Send Messages**
-   - Type message (max 500 characters)
-   - Click "Send Message"
-   - Message appears in streamer's Twitch chat
-   - Uses viewer's own Twitch account
-
-3. **Rate Limiting**
-   - 20 messages per minute
-   - 100 messages per hour
-   - Visual feedback when limit exceeded
-
-4. **Security**
-   - Separate JWT tokens for viewers vs streamers
-   - Tokens encrypted in database
-   - All messages logged to `viewer_message_history`
-   - CSRF protection with state parameter
-
-## 📋 Post-Configuration Testing Checklist
-
-Once the Twitch redirect URI is added:
-
-- [ ] Visit `https://allch.at/chat/caesarlp`
-- [ ] Click "Login with Twitch"
-- [ ] Complete OAuth authorization
-- [ ] Verify redirected back to chat page
-- [ ] Verify "Logged in as {username}" appears
-- [ ] Type a test message
-- [ ] Click "Send Message"
-- [ ] Verify message appears in caesarlp's Twitch chat
-- [ ] Send 21 messages rapidly to test rate limiting
-- [ ] Verify rate limit error appears
-- [ ] Click logout
-- [ ] Verify session cleared
-
-## 🔍 Monitoring
-
-### Database Queries
-
-Check viewer sessions:
-```sql
-SELECT username, platform, created_at
-FROM viewer_sessions
-ORDER BY created_at DESC
-LIMIT 10;
-```
-
-Check message history:
-```sql
-SELECT vs.username, vml.message_text, vml.sent_at, vml.success
-FROM viewer_message_history vml
-JOIN viewer_sessions vs ON vml.viewer_session_id = vs.id
-ORDER BY vml.sent_at DESC
-LIMIT 20;
-```
-
-### Kubernetes Logs
-
-```bash
-# Auth service logs (for OAuth and message sending)
-kubectl logs -n allchat -l app=auth-service --tail=100 -f
-
-# API Gateway logs (for proxying)
-kubectl logs -n allchat -l app=api-gateway --tail=100 -f
-```
-
-## 📝 Implementation Summary
-
-### Commits
-- `ea111a424` - Backend message sending implementation
-- `d24ccfad9` - Frontend viewer chat interface
-
-### Files Created (14 total)
-**Backend (8)**:
-- migrations/011_viewer_authentication.sql
-- services/auth-service/models/viewer.go
-- services/auth-service/repository/viewer_repository.go
-- services/auth-service/oauth/viewer_twitch.go
-- services/auth-service/handlers/viewer_auth.go
-- services/auth-service/handlers/chat_send.go
-- services/auth-service/handlers/streamer_info.go
-- (+ modifications to main.go files)
-
-**Frontend (6)**:
-- src/lib/types/viewer.ts
-- src/lib/api/viewer.ts
-- src/lib/stores/viewer-auth-store.ts
-- src/app/chat/[streamer]/page.tsx
-- src/app/chat/auth-success/page.tsx
-- src/app/chat/auth-error/page.tsx
-
----
-
-**Next Action Required**: Add `https://allch.at/api/v1/auth/viewer/twitch/callback` to Twitch app redirect URIs at https://dev.twitch.tv/console/apps
