+++ frontend/src/lib/types/viewer.ts
@@ -37,44 +37,3 @@ export interface ViewerAuthResponse {
   expires_in: number
   viewer_info: ViewerInfo
 }
-
-export interface ViewerLoginResponse {
-  auth_url: string
-}
-
-export interface StreamerInfo {
-  username: string
-  display_name: string
-  platforms: StreamerPlatform[]
-  overlay_id?: string
-}
-
-export interface StreamerPlatform {
-  platform: 'twitch' | 'youtube' | 'kick' | 'tiktok'
-  channel_id: string
-  channel_name: string
-  is_active: boolean
-}
-
-export interface SendMessageRequest {
-  streamer_username: string
-  message: string
-  platform?: string
-}
-
-export interface StreamerSendMessageRequest {
-  message: string
-  platform: string
-}
-
-export interface SendMessageResponse {
-  success: boolean
-  message: string
-}
-
-export interface ViewerAuthState {
-  viewerInfo: ViewerInfo | null
-  viewerToken: string | null
-  loading: boolean
-  streamer: string | null
-}
